convert POC test into e2e test for default tracer and calltracer
