+++ test/e2e/debug_test.go
@@ -733,6 +733,58 @@ func TestDebugTraceBlock(t *testing.T) {
 	}
 }
 
+func Test_DebugFirstBatch(t *testing.T) {
+	if testing.Short() {
+		t.Skip()
+	}
+
+	const l2NetworkURL = "http://localhost:8124"
+	const l2ExplorerRPCComponentName = "l2-explorer-json-rpc"
+
+	var err error
+	if !dockersArePreLaunchedForDebugTests {
+		err = operations.Teardown()
+		require.NoError(t, err)
+
+		defer func() {
+			require.NoError(t, operations.Teardown())
+			require.NoError(t, operations.StopComponent(l2ExplorerRPCComponentName))
+		}()
+	}
+
+	ctx := context.Background()
+	opsCfg := operations.GetDefaultOperationsConfig()
+	if !dockersArePreLaunchedForDebugTests {
+		opsMan, err := operations.NewManager(ctx, opsCfg)
+		require.NoError(t, err)
+		err = opsMan.Setup()
+		require.NoError(t, err)
+
+		err = operations.StartComponent(l2ExplorerRPCComponentName, func() (bool, error) { return operations.NodeUpCondition(l2NetworkURL) })
+		require.NoError(t, err)
+	} else {
+		log.Info("Using pre-launched dockers: no reset Database")
+	}
+
+	debugOptions := map[string]interface{}{
+		"tracer": "callTracer",
+		"tracerConfig": map[string]interface{}{
+			"onlyTopCall": false,
+			"withLog":     true,
+		},
+	}
+
+	response, err := client.JSONRPCCall(l2NetworkURL, "debug_traceBlockByNumber", "0x1", debugOptions)
+	require.NoError(t, err)
+	require.Nil(t, response.Error)
+	require.NotNil(t, response.Result)
+
+	response, err = client.JSONRPCCall(l2NetworkURL, "debug_traceBlockByNumber", "0x1")
+	require.NoError(t, err)
+	require.Nil(t, response.Error)
+	require.NotNil(t, response.Result)
+}
+
 func getTxInResponseDebugTest(t *testing.T, response json.RawMessage, txIndex uint, debugPrefix string) map[string]interface{} {
 	valueMap := []interface{}{}
 	err := json.Unmarshal(response, &valueMap)
